feat(billing): move services list to its dedicated route
ref: #MANAGER-17700
