+++ packages/manager/modules/new-billing/src/autoRenew/actions/warnPendingDebt/billing-autoRenew-warnPendingDebt.html
@@ -1,5 +1,5 @@
 <oui-modal
-    data-heading="{{ ::'autorenew_service_debt_before_paying_title' | translate }}"
+    data-heading="{{ ::'autorenew_service_debt_warn_title' | translate }}"
     data-primary-action="$ctrl.payDebt()"
     data-primary-label="{{:: 'autorenew_service_debt_warn_settle' | translate }}"
     data-secondary-action="$ctrl.goBack()"
