+++ packages/manager/modules/new-billing/src/autoRenew/autorenew.component.js
@@ -3,41 +3,14 @@ import template from './autorenew.html';
 
 export default {
   bindings: {
-    activationLink: '<',
     agreementsLink: '<',
-    billingServices: '<',
-    canDisableAllDomains: '<',
     currentActiveLink: '<',
-    currentUser: '<',
-    defaultPaymentMean: '<',
-    disableAutorenewForDomains: '<',
-    disableBulkAutorenew: '<',
-    enableBulkAutorenew: '<',
-    filters: '<',
-    getServices: '<',
     goToAutorenew: '<',
-    hasAutoRenew: '<',
     homeLink: '<',
-    canDeleteVrack: '<',
     isAutorenewManagementAvailable: '<',
-    isAutorenew2016DeploymentBannerAvailable: '<',
-    isEnterpriseCustomer: '<',
-    nicBilling: '<',
-    nicRenew: '<',
-    nics: '<',
-    offset: '<',
-    onListParamChanges: '<',
-    pageNumber: '<',
-    pageSize: '<',
-    searchText: '<',
-    selectedType: '<',
-    services: '<',
-    serviceTypes: '<',
-    sort: '<',
     sshLink: '<',
     guides: '<',
     trackClick: '<',
-    choiceRenewDayTooltipAvailable: '<',
   },
   controller,
   template,
