+++ packages/manager/modules/new-billing/src/autoRenew/autorenew.module.js
@@ -35,6 +35,7 @@ import update from './actions/update/update.module';
 import terminateVrack from './actions/terminate-vrack';
 import warnNicBilling from './actions/warnNicBilling/warnNicBilling.module';
 import warnPendingDebt from './actions/warnPendingDebt/pending-debt.module';
+import services from './services/services.module';
 
 import component from './autorenew.component';
 import routing from './autorenew.routing';
@@ -78,6 +79,7 @@ angular
     update,
     warnNicBilling,
     warnPendingDebt,
+    services,
   ])
   .config(routing)
   .component('autoRenew', component)
