+++ packages/manager/modules/new-billing/src/autoRenew/autorenew.routing.js
@@ -1,12 +1,6 @@
 import assign from 'lodash/assign';
-import get from 'lodash/get';
-import map from 'lodash/map';
-import range from 'lodash/range';
-
-import { BillingService } from '@ovh-ux/manager-models';
 
 import {
-  NIC_ALL,
   TRACKING_AUTORENEW_PAGE_NAME,
   TRACKING_PAGE_CATEGORY,
 } from './autorenew.constants';
@@ -60,11 +54,22 @@ export default /* @ngInject */ ($stateProvider, coreConfigProvider) => {
         sshLink: /* @ngInject */ ($state) =>
           $state.href('billing.autorenew.ssh', {}, { inherit: false }),
         queryParameters: /* @ngInject */ ($transition$, BillingAutoRenew) => {
-          if ($transition$.to()?.name === 'billing.autorenew') {
+          if ($transition$.to()?.name === 'billing.autorenew.services') {
             BillingAutoRenew.setQueryParams($transition$.params());
           }
           return BillingAutoRenew.getQueryParams();
         },
+        featureAvailability: /* @ngInject */ (ovhFeatureFlipping) =>
+          ovhFeatureFlipping.checkFeatureAvailability([
+            'vrack:delete',
+            'billing:management',
+            'billing:autorenew2016Deployment',
+          ]),
+        isAutorenewManagementAvailable: /* @ngInject */ (featureAvailability) =>
+          featureAvailability?.isFeatureAvailable('billing:management') ||
+          false,
+        hideBreadcrumb: /* @ngInject */ () => true,
+        trackingPrefix: () => 'dedicated::account::billing::autorenew',
         goToAutorenew: /* @ngInject */ (
           $state,
           $timeout,
@@ -89,65 +94,22 @@ export default /* @ngInject */ ($stateProvider, coreConfigProvider) => {
 
           return promise;
         },
-        featureAvailability: /* @ngInject */ (ovhFeatureFlipping) =>
-          ovhFeatureFlipping.checkFeatureAvailability([
-            'vrack:delete',
-            'billing:management',
-            'billing:autorenew2016Deployment',
-          ]),
-        isAutorenewManagementAvailable: /* @ngInject */ (featureAvailability) =>
-          featureAvailability?.isFeatureAvailable('billing:management') ||
-          false,
-        isAutorenew2016DeploymentBannerAvailable: /* @ngInject */ (
-          featureAvailability,
-        ) =>
-          featureAvailability?.isFeatureAvailable(
-            'billing:autorenew2016Deployment',
-          ) || false,
-        canDeleteVrack: /* @ngInject */ (featureAvailability) =>
-          featureAvailability?.isFeatureAvailable('vrack:delete') || false,
-        hideBreadcrumb: /* @ngInject */ () => true,
-        trackingPrefix: () => 'dedicated::account::billing::autorenew',
       },
       !coreConfigProvider.isRegion('US')
         ? {
-            activationLink: /* @ngInject */ ($state) =>
-              $state.href('billing.autorenew.activation'),
             agreementsLink: /* @ngInject */ ($state) =>
               $state.href(
                 'billing.autorenew.agreements',
                 {},
                 { inherit: false },
               ),
-            billingServices: /* @ngInject */ (services) =>
-              map(
-                services.list.results,
-                (service) => new BillingService(service),
-              ),
-            canDisableAllDomains: /* @ngInject */ (services) =>
-              services.bulkDomains,
-            /* @ngInject */
-            defaultPaymentMean: (ovhPaymentMethod) =>
-              ovhPaymentMethod.getDefaultPaymentMethod(),
-            disableAutorenewForDomains: /* @ngInject */ ($state) => () =>
-              $state.go('billing.autorenew.disableDomainsBulk'),
-            disableBulkAutorenew: /* @ngInject */ ($state) => (services) =>
-              $state.go('billing.autorenew.disable', {
-                services: map(services, 'id').join(','),
-              }),
 
-            enableBulkAutorenew: /* @ngInject */ ($state) => (services) =>
-              $state.go('billing.autorenew.enable', {
-                services: map(services, 'id').join(','),
-              }),
-            endStrategies: /* @ngInject */ (endStrategyEnum) =>
-              endStrategyEnum.reduce(
-                (object, strategy) => ({
-                  ...object,
-                  [strategy]: strategy,
-                }),
-                {},
-              ),
+            homeLink: /* @ngInject */ ($state) =>
+              $state.href('billing.autorenew.services', {}, { inherit: false }),
+
+            breadcrumb: /* @ngInject */ ($translate) =>
+              $translate.instant('billing_title'),
+
             endStrategyEnum: /* @ngInject */ ($http) =>
               $http
                 .get('/services.json')
@@ -156,89 +118,6 @@ export default /* @ngInject */ ($stateProvider, coreConfigProvider) => {
                     data.models['services.billing.engagement.EndStrategyEnum']
                       ?.enum,
                 ),
-            filters: /* @ngInject */ ($transition$) =>
-              JSON.parse($transition$.params().filters),
-            isEnterpriseCustomer: /* @ngInject */ (currentUser) =>
-              currentUser.isEnterprise,
-
-            hasAutoRenew: /* @ngInject */ (billingRenewHelper) => (service) =>
-              billingRenewHelper.serviceHasAutomaticRenew(service),
-
-            homeLink: /* @ngInject */ ($state) =>
-              $state.href('billing.autorenew', {}, { inherit: false }),
-
-            nicBilling: /* @ngInject */ ($transition$) =>
-              $transition$.params().nicBilling,
-            nicRenew: /* @ngInject */ (BillingAutoRenew, services) =>
-              BillingAutoRenew.getAutorenew().then((nicRenew) => ({
-                ...nicRenew,
-                isMandatory: services.userMustApproveAutoRenew,
-                renewDays: range(1, 30),
-              })),
-
-            nics: /* @ngInject */ ($translate, services) => [
-              $translate.instant(NIC_ALL),
-              ...get(services, 'nicBilling', []),
-            ],
-
-            offset: /* @ngInject */ (pageNumber, pageSize) =>
-              pageSize * (pageNumber - 1),
-
-            onListParamChanges: /* @ngInject */ ($state) => (params) =>
-              $state.go('.', params, { notify: false }),
-
-            pageNumber: /* @ngInject */ ($transition$) =>
-              parseInt($transition$.params().pageNumber, 10),
-            pageSize: /* @ngInject */ ($transition$) =>
-              parseInt($transition$.params().pageSize, 10),
-
-            searchText: /* @ngInject */ ($transition$) =>
-              $transition$.params().searchText,
-
-            selectedType: /* @ngInject */ ($transition$) =>
-              $transition$.params().selectedType,
-
-            services: /* @ngInject */ (
-              BillingAutoRenew,
-              filters,
-              nicBilling,
-              pageSize,
-              offset,
-              searchText,
-              selectedType,
-              sort,
-            ) =>
-              BillingAutoRenew.getServices(
-                pageSize,
-                offset,
-                searchText,
-                selectedType,
-                filters.expiration,
-                filters.status,
-                filters.state,
-                sort,
-                nicBilling,
-              ),
-
-            choiceRenewDayTooltipAvailable: /* @ngInject */ (
-              ovhFeatureFlipping,
-            ) =>
-              ovhFeatureFlipping
-                .checkFeatureAvailability(['billing:choiceRenewDayTooltip'])
-                .then((commitmentAvailability) =>
-                  commitmentAvailability?.isFeatureAvailable(
-                    'billing:choiceRenewDayTooltip',
-                  ),
-                ),
-
-            serviceTypes: /* @ngInject */ (BillingAutoRenew, services) =>
-              BillingAutoRenew.getServicesTypes(services),
-
-            sort: /* @ngInject */ ($transition$) =>
-              JSON.parse($transition$.params().sort),
-
-            breadcrumb: /* @ngInject */ ($translate) =>
-              $translate.instant('billing_title'),
           }
         : {},
     ),
@@ -255,10 +134,14 @@ export default /* @ngInject */ ($stateProvider, coreConfigProvider) => {
       transition
         .injector()
         .getAsync('isAutorenewManagementAvailable')
-        .then(
-          (isAutorenewManagementAvailable) =>
-            !isAutorenewManagementAvailable && 'billing.autorenew.ssh',
-        ),
+        .then((isAutorenewManagementAvailable) => {
+          if (!isAutorenewManagementAvailable) {
+            return 'billing.autorenew.ssh';
+          }
+          return transition.to()?.name === 'billing.autorenew'
+            ? 'billing.autorenew.services'
+            : null;
+        }),
   });
 
   $stateProvider.state('billing.autorenew.service', {
