+++ packages/manager/modules/new-billing/src/autoRenew/services/services.component.js
@@ -1,0 +1,39 @@
+import controller from './services.controller';
+import template from './services.html';
+
+export default {
+  bindings: {
+    activationLink: '<',
+    billingServices: '<',
+    canDisableAllDomains: '<',
+    currentUser: '<',
+    defaultPaymentMean: '<',
+    disableAutorenewForDomains: '<',
+    disableBulkAutorenew: '<',
+    enableBulkAutorenew: '<',
+    filters: '<',
+    getServices: '<',
+    goToAutorenew: '<',
+    hasAutoRenew: '<',
+    canDeleteVrack: '<',
+    isAutorenewManagementAvailable: '<',
+    isAutorenew2016DeploymentBannerAvailable: '<',
+    isEnterpriseCustomer: '<',
+    nicBilling: '<',
+    nicRenew: '<',
+    nics: '<',
+    offset: '<',
+    onListParamChanges: '<',
+    pageNumber: '<',
+    pageSize: '<',
+    searchText: '<',
+    selectedType: '<',
+    services: '<',
+    serviceTypes: '<',
+    sort: '<',
+    trackClick: '<',
+    choiceRenewDayTooltipAvailable: '<',
+  },
+  controller,
+  template,
+};
