+++ packages/manager/modules/new-billing/src/autoRenew/services/services.module.js
@@ -1,0 +1,13 @@
+import angular from 'angular';
+
+import component from './services.component';
+import routing from './services.routing';
+
+const moduleName = 'ovhManagerBillingAutorenewServices';
+
+angular
+  .module(moduleName, [])
+  .config(routing)
+  .component('services', component);
+
+export default moduleName;
